+++ packages/riverpod/test/old/legacy_providers/deprecated/state_notifier_provider/auto_dispose_family_test.dart
@@ -1,7 +1,7 @@
 import 'package:riverpod/riverpod.dart';
 import 'package:test/test.dart';
 
-import '../../utils.dart';
+import '../../../utils.dart';
 
 void main() {
   group('StateNotifier.family', () {
