+++ packages/riverpod/test/old/legacy_providers/deprecated/state_notifier_provider/auto_dispose_state_notifier_provider_test.dart
@@ -2,7 +2,7 @@ import 'package:mockito/mockito.dart';
 import 'package:riverpod/riverpod.dart';
 import 'package:test/test.dart';
 
-import '../../utils.dart';
+import '../../../utils.dart';
 
 void main() {
   test('can read and set current StateNotifier', () async {
