+++ packages/riverpod/test/old/legacy_providers/deprecated/state_notifier_provider/family_test.dart
@@ -2,7 +2,7 @@ import 'package:mockito/mockito.dart';
 import 'package:riverpod/riverpod.dart';
 import 'package:test/test.dart';
 
-import '../../utils.dart';
+import '../../../utils.dart';
 import 'state_notifier_provider_test.dart';
 
 void main() {
