+++ packages/riverpod/test/old/legacy_providers/deprecated/state_notifier_provider/state_notifier_provider_test.dart
@@ -4,7 +4,7 @@ import 'package:mockito/mockito.dart';
 import 'package:riverpod/riverpod.dart';
 import 'package:test/test.dart';
 
-import '../../utils.dart';
+import '../../../utils.dart';
 
 void main() {
   test('supports overrideWith', () {
